refactor: heap and use scope for stable again
