+++ src/console/src/factory/mission_control.rs
@@ -1,7 +1,7 @@
 use crate::constants::FREEZING_THRESHOLD_ONE_YEAR;
 use crate::controllers::update_mission_control_controllers;
 use crate::store::heap::increment_mission_controls_rate;
-use crate::store::{
+use crate::store::stable::{
     add_mission_control, delete_mission_control, get_mission_control, init_empty_mission_control,
 };
 use crate::types::state::MissionControl;
