+++ src/console/src/store/heap/mod.rs
@@ -1,0 +1,11 @@
+mod controllers;
+mod fees;
+mod invitation_codes;
+mod rates;
+mod releases_metadata;
+
+pub use controllers::*;
+pub use fees::*;
+pub use invitation_codes::*;
+pub use rates::*;
+pub use releases_metadata::*;
