+++ src/console/src/store/mod.rs
@@ -1,5 +1,3 @@
 pub mod heap;
 mod services;
-mod stable;
-
-pub use stable::*;
+pub mod stable;
